requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
Pillow
python-docx
docx2pdf
pytesseract